+++ media/libstagefright/FileSource.cpp
@@ -111,7 +111,7 @@ ssize_t FileSource::readAt(off64_t offset, void *data, size_t size) {
    } else {
         off64_t result = lseek64(mFd, offset + mOffset, SEEK_SET);
         if (result == -1) {
-            ALOGE("seek to %lld failed", offset + mOffset);
+            ALOGE("seek to %lld failed", (long long)(offset + mOffset));
             return UNKNOWN_ERROR;
         }
